+++ ghost/core/core/frontend/services/assets-minification/assets-minification-base.js
