+++ ghost/core/core/frontend/services/assets-minification/index.js
@@ -1,5 +1,5 @@
-const CardAssets = require('./CardAssets');
-const AdminAuthAssets = require('./AdminAuthAssets');
+const CardAssets = require('./card-assets');
+const AdminAuthAssets = require('./admin-auth-assets');
 const cardAssets = new CardAssets();
 const adminAuthAssets = new AdminAuthAssets();
