+++ ghost/core/core/frontend/services/routing/email-router.js
@@ -1,18 +1,18 @@
-const ParentRouter = require('./ParentRouter');
+const ParentRouter = require('./parent-router');
 const urlUtils = require('../../../shared/url-utils');
 const controllers = require('./controllers');
 
 /**
- * @description Preview Router.
+ * @description Email Router.
  */
-class PreviewRouter extends ParentRouter {
+class EmailRouter extends ParentRouter {
     constructor(RESOURCE_CONFIG) {
-        super('PreviewRouter');
+        super('EmailRouter');
 
-        this.RESOURCE_CONFIG = RESOURCE_CONFIG.QUERY.previews;
+        this.RESOURCE_CONFIG = RESOURCE_CONFIG.QUERY.email;
 
         // @NOTE: hardcoded, not configurable
-        this.route = {value: '/p/'};
+        this.route = {value: '/email/'};
 
         this._registerRoutes();
     }
@@ -25,8 +25,8 @@ class PreviewRouter extends ParentRouter {
         // REGISTER: prepare context
         this.router().use(this._prepareContext.bind(this));
 
-        // REGISTER: actual previews route
-        this.mountRoute(urlUtils.urlJoin(this.route.value, ':uuid', ':options?'), controllers.previews);
+        // REGISTER: actual email route
+        this.mountRoute(urlUtils.urlJoin(this.route.value, ':uuid', ':options?'), controllers.email);
     }
 
     /**
@@ -40,11 +40,11 @@ class PreviewRouter extends ParentRouter {
         res.routerOptions = {
             type: 'entry',
             query: this.RESOURCE_CONFIG,
-            context: ['preview']
+            context: ['emailPost']
         };
 
         next();
     }
 }
 
-module.exports = PreviewRouter;
+module.exports = EmailRouter;
