+++ ghost/core/core/frontend/services/routing/static-pages-router.js
@@ -1,6 +1,6 @@
 const debug = require('@tryghost/debug')('routing:static-pages-router');
 const urlUtils = require('../../../shared/url-utils');
-const ParentRouter = require('./ParentRouter');
+const ParentRouter = require('./parent-router');
 const controllers = require('./controllers');
 
 /**
