+++ ghost/core/core/frontend/services/routing/taxonomy-router.js
@@ -1,7 +1,7 @@
 const debug = require('@tryghost/debug')('routing:taxonomy-router');
 const config = require('../../../shared/config');
-const ParentRouter = require('./ParentRouter');
-const RSSRouter = require('./RSSRouter');
+const ParentRouter = require('./parent-router');
+const RSSRouter = require('./rss-router');
 const urlUtils = require('../../../shared/url-utils');
 const controllers = require('./controllers');
 const middleware = require('./middleware');
