+++ ghost/core/core/frontend/services/routing/unsubscribe-router.js
@@ -1,4 +1,4 @@
-const ParentRouter = require('./ParentRouter');
+const ParentRouter = require('./parent-router');
 const controllers = require('./controllers');
 
 /**
