+++ ghost/core/core/frontend/services/sitemap/base-site-map-generator.js
