+++ ghost/core/core/frontend/services/sitemap/page-map-generator.js
@@ -1,5 +1,5 @@
 const _ = require('lodash');
-const BaseMapGenerator = require('./BaseSiteMapGenerator');
+const BaseMapGenerator = require('./base-site-map-generator');
 
 class PageMapGenerator extends BaseMapGenerator {
     constructor(opts) {
