+++ ghost/core/core/frontend/services/sitemap/site-map-index-generator.js
