+++ ghost/core/core/frontend/web/routes.js
@@ -3,7 +3,7 @@ const routing = require('../services/routing');
 
 /**
  *
- * @param {import('../services/routing/RouterManager').RouterConfig} routerConfig
+ * @param {import('../services/routing/router-manager').RouterConfig} routerConfig
  * @returns {import('express').Router}
  */
 module.exports = function siteRoutes(routerConfig) {
