+++ ghost/core/test/unit/frontend/services/routing/bootstrap.test.js
@@ -1,7 +1,7 @@
 const should = require('should');
 const sinon = require('sinon');
-const CollectionRouter = require('../../../../../core/frontend/services/routing/CollectionRouter');
-const RouterManager = require('../../../../../core/frontend/services/routing/RouterManager');
+const CollectionRouter = require('../../../../../core/frontend/services/routing/collection-router');
+const RouterManager = require('../../../../../core/frontend/services/routing/router-manager');
 const registry = require('../../../../../core/frontend/services/routing/registry');
 
 const RESOURCE_CONFIG = {QUERY: {post: {controller: 'posts', resource: 'posts'}}};
