+++ ghost/core/test/unit/frontend/services/theme-engine/i18next.test.js
@@ -1,6 +1,6 @@
 const should = require('should');
 const sinon = require('sinon');
-const ThemeI18n = require('../../../../../core/frontend/services/theme-engine/i18next/ThemeI18n');
+const ThemeI18n = require('../../../../../core/frontend/services/theme-engine/i18next/theme-i18n');
 const path = require('path');
 
 describe('NEW i18nextThemeI18n Class behavior', function () {
@@ -23,7 +23,7 @@ describe('NEW i18nextThemeI18n Class behavior', function () {
         i18n.init({activeTheme: 'locale-theme', locale: 'fr'});
         i18n._locale.should.eql('fr');
     });
-    
+
     it('initializes with theme path', function () {
         i18n.init({activeTheme: 'locale-theme', locale: 'de'});
         const result = i18n.t('Top left Button');
